Address review comments

+++ model/modeldecoder/rumv3/transaction_test.go
@@ -382,7 +382,7 @@ func TestDecodeMapToTransactionModel(t *testing.T) {
 		var input transaction
 		input.Context.Tags = common.MapStr{
 			"a": "b",
-			"c": 12315124131,
+			"c": float64(12315124131),
 			"d": 12315124131.12315124131,
 			"e": true,
 		}
