+++ model/modeldecoder/v2/span_test.go
@@ -449,7 +449,7 @@ func TestDecodeMapToSpanModel(t *testing.T) {
 		var input span
 		input.Context.Tags = common.MapStr{
 			"a": "b",
-			"c": 12315124131,
+			"c": float64(12315124131),
 			"d": 12315124131.12315124131,
 			"e": true,
 		}
